+++ backend/src/main/java/com/summerHack/diningTogether/controller/FoodController.java
@@ -32,7 +32,7 @@ public class FoodController {
 
     @ApiOperation(value = "show food by id")
     @GetMapping("/{id}")
-    public Food getFood(@PathVariable("id") int id) {
+    public Food getFood(@PathVariable("id") long id) {
         return this.foodService.getFoodById(id);
     }
 
@@ -44,18 +44,18 @@ public class FoodController {
 
     @PutMapping("/{id}")
     @ApiOperation(value = "modify the information of a meal")
-    public Food modifyFood(@PathVariable("id") int id, @RequestBody Food food) {
+    public Food modifyFood(@PathVariable("id") long id, @RequestBody Food food) {
         return this.foodService.updateFood(id, food);
     }
 
     @DeleteMapping("/{id}")
     @ApiOperation(value = "delete a meal")
-    public Food deleteFood(@PathVariable("id") Integer id) {
+    public Food deleteFood(@PathVariable("id") long id) {
         return this.foodService.deleteById(id);
     }
 
     @PutMapping("/{id}/confirm")
-    public Food confirmFood(@PathVariable("id") Integer id) {
+    public Food confirmFood(@PathVariable("id") long id) {
         return this.foodService.confirmFood(id);
     }
 }
